retsnoop: emit deepest success stack trace if requested
The most complete *success* stack trace is exactly at the time when the
deepest successful function invocation returns. So emit full stack trace at
that point, if success stack traces are requested.

We should clean up duplicate stack trace emitting later on, but for now this
is a useful investigation & exploration functionality.

Signed-off-by: Andrii Nakryiko <andrii@kernel.org>