mass_attacher: try addrs-based multi kprobe and fallback to syms-based
Addrs-based multi kprobe attachment is the fastest but is more strict in
terms of handling of notrace functions (and probably some other
restrictions). Syms-based seems to be more permissive. So attempt
addrs-based multi-attach and fallback to slower syms-based one (which is
still much faster than one-by-one kprobe attachment) to get the best
possible attach performance.

Signed-off-by: Andrii Nakryiko <andrii@kernel.org>