retsnoop: allow error values allow/deny filter
Allow to filter out "uninteresting" error stacks by specifying allowed
errors with `-x Exxx` and/or errors that should be ignored with `-X Exxx`.

When -x is used, only stacks that contains *at least one* of the allowed
errors will be reported. If not `-x` argument is ever specified, all
errors are allowed, except those explicitly denied with `-X`.

When -X is used, any stack that contains *even one* of the blacklisted
errors will be ignored.

NULL is considered an error as well, so `-x NULL` can be used to only
show stacks that return NULL in one of the captured functions. `-X NULL`
can be used to ignore any such stacks.

Signed-off-by: Andrii Nakryiko <andrii@kernel.org>