retsnoop: default lbr mode to any_return by default
any_return seems to be a better default that allows to capture more
function calls given a limited amount of LBR records available. Previous
default any is useful when trying to peer into last non-inlined
function, but to get there one usually needs to find that last function,
which any_return helps with.

Signed-off-by: Andrii Nakryiko <andrii@kernel.org>