mass_attach: add debugging of failed multi-kprobe attachment
Add `--debug multi-kprobe` option, which performs binary search of all
the functions that retsnoop fails to attach to with multi-kprobe
attachments.

This is very useful to know what functions to filter out, if something
can't be attached. It also is a great way to debug various bugs in
retsnoop as related to figuring out allowed kprobes.

Signed-off-by: Andrii Nakryiko <andrii@kernel.org>