ci: Use dedicated GitHub Action to package sources for release workflow
Instead of re-implementing source packaging for multiple projects, we've
come up with a dedicated GitHub Action. Let's use it for the release
workflow.

This Action should take into account the export-ignore rules from
.gitattributes, modulo some adjustment that was just merged into libbpf
and has not been propagated to retsnoop's submodules yet. In the
meantime, this means that we have no difference with the previous
behaviour, where the release workflow would ignore these exclusion rules
anyway.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>