ci: release: Cross-compile addr2line outside of Docker container
Although the release workflow used to succeed, the arm64 is now failing
because the update for Cargo's registry index gets OOM-killed, when
happening in the Docker container we use for cross-compiling. I'm not
sure what caused that change, but this is something we want to address.

I tested several options: using the "sparse" protocol to download less
data for the index, or using the git CLI to update it ("cargo update
--config net.git-fetch-with-cli=true"), which is supposed to require
less memory. None of these solutions was enough. I managed to fix the
workflow by: 1) running "cargo update" ouside of the container, 2)
mounting Cargo's home directory into the container:

  docker run --platform linux/arm64 --rm -v $(pwd):/build \
      -v ${HOME}/.cargo/git:/usr/local/cargo/git \
      -v ${HOME}/.cargo/registry:/usr/local/cargo/registry \
      ubuntu:22.04 \
      ...

and 3) running another "cargo update" inside of the container (using
sparse and git CLI for both updates).

This solution seems heavy and not optimal, and the job would still take
around 20 minutes to complete.

Instead, we can build addr2line outside of the container, directly on
the runner. This is not super-clean in the sense that part of retsnoop
is cross-compiled on the host, and another part of it inside of the
container; also, we set "CARGO=true" in the container to avoid
re-triggering a build. But it works, and it's efficient: the job now
lasts about 6 minutes.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>