bpftool: Pass additional compile flags as EXTRA_CFLAGS, not CFLAGS
Bpftool expects users to pass additional compilation flags via
EXTRA_CFLAGS, not via CFLAGS. This is because when compiled from the
kernel repo, CFLAGS are usually overwritten by the build system.

When we build retsnoop with CFLAGS=--static, the static flag is passed
down to bpftool's Makefile. Then we hit an issue: --static is ignored
for feature probing at build time, but taken into account for the build
itself. Under certain conditions this results in a discrepency and
bpftool's Makefile believes LLVM support is available (given it tested
without the --static flag) whereas it is not set up for static builds.

Let's address this by passing retsnoop's CFLAGS via bpftool's
EXTRA_CFLAGS, and resetting its CFLAGS. We must pass the CFLAGS to the
"make" invocation through the environment, and not as a command line
argument, to avoid overriding the CFLAGS values that we set inside of
bpftool's Makefile.

Reported-by: Andrii Nakryiko <andrii@kernel.org>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>