retsnoop: normalize func/glob display name output
Add NAME_MOD() macro that formats glob or function spec as either
'<name>' if it's not module-specific or '<name> [<module>]' if it is
module-specific. Normalize all the output code to use this approach and
not have to do ad-hoc work arounds to minimize conditional logic just to
output this information consistently.

Signed-off-by: Andrii Nakryiko <andrii@kernel.org>