globs: abstract away glob set matching logic
This allows to keep reporting and stats counting separate from the (now
reusable) logic of matching a given set of allow/deny globs.

There are 4 possible outcomes for any given glob set. It could allow
(match) or deny (mismatch), and do it either explicitly (there is
a matching allow/deny glob) or implicitly (no glob matched, but match or
mismatch is implicit, depending on whether there is an explicit
GLOB_ALLOW glob).

Signed-off-by: Andrii Nakryiko <andrii@kernel.org>