logic: override BTF info for BPF helpers
BPF helpers are defined as returning u64 in kernel, and so that's what
BTF records. But in reality there is often either pointer being returned
(encoded as u64) or signed integer for which negative value is some sort
of error code.

So instead of relying on BTF information and claim that none of BPF
helpers can fail, encode entire 211 BPF helpers (and this sets is not
going to change going forward, so no ongoing maintenance is required)
and their real function flags according to UAPI documentation.

Signed-off-by: Andrii Nakryiko <andrii@kernel.org>