mass_attacher: work around kernel's objpool -ENOMEM bug for multi-kprobes
Kernel versions 6.7 through 6.11 are affected by a memory allocation bug
in objpool code, which is used for multi-kretprobes. As such, when
sufficiently large number of kernel functions are being attached to with
multi-kretprobe, we can get -ENOMEM even though the system has plenty of
memory.

Work around this by doing multi-batched attachment and dynamically
reduce the batch size if we keep getting -ENOMEM. We do this batch size
adjustment only for kretprobe, and also switch the order of attachment
to try kretprobe attachment before kprobe attachment to simplify the
algorithm and leave the overall code structure similar.

Signed-off-by: Andrii Nakryiko <andrii@kernel.org>